adicionar alunos no alunos.csv e adicionar condicional para pular compilação de certificados já existentes

+++ certificados/gerar_certificado.R
@@ -13,32 +13,36 @@
 gerar_certificado <-
   function(modelo,
            nome_aluno,
-           nome_evento,
-           data_evento,
-           local_evento,
            saida_pdf,
+           caminho_pdf,
            diretorio_rmd) {
-    cat("\n Iniciando a compilação:", saida_pdf, "\n")
-    
-    # Criar um arquivo qmd temporário com aluno e informações do evento
-    modelo_certificado <- readr::read_file(modelo)
-    
-    # Modificar NOMEALUNO pelo nome do aluno para o modelo
-    # temporário de cada iteração
-    modelo_tmp <- modelo_certificado |> 
-      stringr::str_replace("NOMEALUNO", nome_aluno)
-    
-    # O diretorio_rmd tem que ser definido para o
-    # rmarkdown::render() saber onde salvar o arquivo temporário
-    arquivo_rmd <- tempfile(tmpdir = diretorio_rmd, fileext = ".Rmd")
-    readr::write_file(modelo_tmp, arquivo_rmd)
-    
-    # Criar os certificados usando R Markdown
-    rmarkdown::render(arquivo_rmd,
-                      output_file = saida_pdf,
-                      quiet = TRUE)
-    
-    # Arquivo .Rmd temporário pode ser excluído após cada iteração
-    file.remove(arquivo_rmd)
-    cat("\n Finalizado:", saida_pdf, "\n")
+    if (file.exists(caminho_pdf)) {
+      cat("\n Pulando arquivo:", saida_pdf, "\n")
+    }
+    else {
+      cat("\n Iniciando a compilação:", saida_pdf, "\n")
+      
+      # Criar um arquivo qmd temporário com aluno e informações do evento
+      modelo_certificado <- readr::read_file(modelo)
+      
+      # Modificar NOMEALUNO pelo nome do aluno para o modelo
+      # temporário de cada iteração
+      modelo_tmp <- modelo_certificado |>
+        stringr::str_replace("NOMEALUNO", nome_aluno)
+      
+      # O diretorio_rmd tem que ser definido para o
+      # rmarkdown::render() saber onde salvar o arquivo temporário
+      arquivo_rmd <-
+        tempfile(tmpdir = diretorio_rmd, fileext = ".Rmd")
+      readr::write_file(modelo_tmp, arquivo_rmd)
+      
+      # Criar os certificados usando R Markdown
+      rmarkdown::render(arquivo_rmd,
+                        output_file = saida_pdf,
+                        quiet = TRUE)
+      
+      # Arquivo .Rmd temporário pode ser excluído após cada iteração
+      file.remove(arquivo_rmd)
+      cat("\n Finalizado:", saida_pdf, "\n")
+    }
   }
